created TakeLoan and RepayLoan components in NftLoansModal folder, refactored NftLoansModal.jsx and zinarLoanContractService.js, deleted NftLoansModal.css
