+++ src/common/Modals/NftLoansModals/NftLoansModal.jsx
@@ -1,27 +1,15 @@
 import React, {useEffect, useState} from "react";
-import "./NftLoansModal.css";
 import "../NftModals/NftModals.css";
 import CancelOutlinedIcon from "@mui/icons-material/CancelOutlined";
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
-import { ethers } from "ethers";
-import contractABI from "../../../utils/ZinarLoans.json";
-import nftcontractABI from "../../../utils/ZinarNFTtest.json";
-import { fetchNFTsForContract, fetchNativeBalance } from "../../../services/alchemy-sdk";
-import { setIntRate, getAdminFee } from "../../../services/zinarLoanContractService";
-import { setLoanPrincipal } from "../../../services/mintNftContractService";
+import { fetchNFTsForContract } from "../../../services/alchemy-sdk";
+import TakeLoanModalBox from "./NftLoansModalBox/TakeLoan";
 
 const NftLoansModal = (props) => {
-  const LOAN_CONTRACT_ADDRESS = "0xE73207f981F7787170B4fC6C3348D40974974dae";
-  const NFT_CONTRACT_ADDRESS = "0x161ED8dc509bDAE1b7FAaaD5b48269bC7c283c05";
   const [zinarNft, setZinarNft] = useState([]);
   const { setNftLoansModal } = props;
-  const { ethereum } = window;
-  const provider = new ethers.providers.Web3Provider(ethereum);
-  const signer = provider.getSigner();
-  // connect to the contract you want to execute
-  const contract = new ethers.Contract(LOAN_CONTRACT_ADDRESS, contractABI.abi, signer);
 
   const settings = {
     dots: false,
@@ -42,77 +30,6 @@ const NftLoansModal = (props) => {
     getNfts();
   }, []);
 
-  const getApprovalReceipt = async (_nftId) => {
-    const nftContract = new ethers.Contract(NFT_CONTRACT_ADDRESS, nftcontractABI.abi, signer);
-    try { 
-      const approveTx = await nftContract.approve(LOAN_CONTRACT_ADDRESS, _nftId);
-      let receipt = approveTx.wait();
-      return receipt;
-    } catch (error) {
-      console.log(error);
-    }
-  }
-
-  const takeZinarLoan = async(loanAmount, nftId, InterestRate) => {
-    await InterestRate;
-    console.log(InterestRate);
-
-    try {
-      const approvalReceipt = await getApprovalReceipt(nftId);
-      if (!approvalReceipt) {
-        return false;
-      } else if (approvalReceipt.status === 1) {
-        console.log('Accessing wallet to pay gas');
-        const takeLoan = await contract.beginLoan(
-          loanAmount, 
-          nftId, 
-          InterestRate, 
-          NFT_CONTRACT_ADDRESS,
-          {value: getAdminFee()}
-        );
-        const receipt = await takeLoan.wait();
-    
-        if (receipt.status === 1) {
-          alert("Loan started! https://mumbai.polygonscan.com/tx/"+takeLoan.hash);
-          
-        } else {
-          alert("Transaction failed! Please try again");
-        }
-      } else {
-        return false;
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  }
-
-  const beginLoan = async(nft = {}) => {
-    let selectedNft = {...nft};
-    try {
-      takeZinarLoan(setLoanPrincipal(selectedNft.name), selectedNft.id, setIntRate(selectedNft.name));
-    }catch(error) {
-      console.log(error)
-    }
-  }
-
-  const paybackZinarLoan = async(loanId) => {
-    console.log('Accessing wallet to pay gas');
-    
-    const payLoan = await contract.payBackLoan(
-      loanId,
-      {value: getAdminFee()}
-    );
-    const receipt = await payLoan.wait();
-
-    if (receipt.status === 1) {
-      alert("Loan started! https://mumbai.polygonscan.com/tx/"+payLoan.hash);
-      
-    } else {
-      alert("Transaction failed! Please try again");
-    }
-
-  }
-
   return (
     <div className="modalBackground">
       <div className="modalContainer">
@@ -125,21 +42,12 @@ const NftLoansModal = (props) => {
         <div className="modalContent">
           <Slider {...settings}>
             {zinarNft.map(nft => (
-              <div className="nftLoansModal">
-                <div className="nftLoansModalTitle">{nft.name}</div>
-                <div className="nftLoansModalContent">
-                  <div key={nft.id}>
-                    <video autoPlay loop src={nft.image} width={250} height={250}/>
-                  </div>
-                  <div>
-                    <div>BUSD Balance: </div>
-                    <div>NFT ID: {nft.id}</div>
-                    <div>Loaned Amount: </div>
-                    <div>Accrued Interest</div>
-                    <button onClick={() => beginLoan(nft)}>Borrow Now</button>
-                  </div>
-                </div>
-              </div>
+              <TakeLoanModalBox
+                key={nft}
+                nftName={nft.name}
+                nftImage={nft.image}
+                nftId={nft.id}
+              />
             ))}
           </Slider>
         </div>
