+++ src/common/Modals/NftLoansModals/NftLoansModalBox/NftLoansModalBox.css
